# Projection performance: Don't allocate MvccData to then immediately overwrite it