--- a/src/benchmarklib/tpch/tpch_table_generator.cpp
+++ b/src/benchmarklib/tpch/tpch_table_generator.cpp
@@ -139,7 +139,14 @@ class TableBuilder {
       vector = std::decay_t<decltype(vector)>();
       vector.reserve(_estimated_rows_per_chunk);
     });
-    _table->append_chunk(segments);
+
+    // Create initial MvccData if MVCC is enabled
+    auto mvcc_data = std::shared_ptr<MvccData>{};
+    if (_use_mvcc == UseMvcc::Yes) {
+      mvcc_data = std::make_shared<MvccData>(segments.front()->size(), CommitID{0});
+    }
+
+    _table->append_chunk(segments, mvcc_data);
   }
 };