--- a/src/lib/import_export/csv_parser.cpp
+++ b/src/lib/import_export/csv_parser.cpp
@@ -84,7 +84,9 @@ std::shared_ptr<Table> CsvParser::parse(const std::string& filename, const std::
   CurrentScheduler::wait_for_tasks(tasks);
 
   for (auto& segments : segments_by_chunks) {
-    table->append_chunk(segments);
+    DebugAssert(!segments.empty(), "Empty chunks shouldn't occur when importing CSV");
+    const auto mvcc_data = std::make_shared<MvccData>(segments.front()->size(), CommitID{0});
+    table->append_chunk(segments, mvcc_data);
   }
 
   return table;