--- a/src/lib/operators/abstract_operator.cpp
+++ b/src/lib/operators/abstract_operator.cpp
@@ -71,8 +71,6 @@ std::shared_ptr<const Table> AbstractOperator::get_output() const {
       }(),
       "Empty chunk returned from operator " + description());
 
-  DebugAssert(!_output || _output->column_count() > 0, "Operator " + description() + " did not output any columns");
-
   return _output;
 }