… resources/test_data/csv/no_columns.csv.json
… src/lib/operators/union_positions.cpp
… src/lib/storage/table.hpp
… src/test/operators/import_csv_test.cpp