Fix #1645 

Also, avoid redundant calls to `input_table_left()` for each `Chunk` to save a shared_ptr copy.
Also, remove `Chunk::set_mvcc_data()` because it isn't needed anymore since `Table::append_chunk()` now takes the new Chunk's MvccData as well. One entry point less for potential concurrency problems.

TPC-H with sf=5 
```
+----------------+-----------------+------+-----------------+------+------------+---------------------------------+
| Benchmark      | prev. iter/s    | runs | new iter/s      | runs | change [%] | p-value (significant if <0.001) |
+----------------+-----------------+------+-----------------+------+------------+---------------------------------+
| TPC-H 1        | 0.0490739420056 | 10   | 0.0489541068673 | 10   | -0%        |                          0.4264 |
| TPC-H 2        | 0.671977341175  | 60   | 0.730327427387  | 60   | +9%        |                          0.0000 |
| TPC-H 3        | 0.448871999979  | 60   | 0.448813527822  | 60   | -0%        |                          0.9500 |
| TPC-H 4        | 0.203322425485  | 41   | 0.208486706018  | 42   | +3%        |                          0.0162 |
| TPC-H 5        | 0.277782648802  | 56   | 0.319226652384  | 60   | +15%       |                          0.0000 |
| TPC-H 6        | 0.775415897369  | 60   | 0.792824447155  | 60   | +2%        |                          0.0666 |
| TPC-H 7        | 0.0973886549473 | 20   | 0.0974526852369 | 20   | +0%        |                          0.9659 |
| TPC-H 8        | 0.429121643305  | 60   | 0.514973104     | 60   | +20%       |                          0.0000 |
| TPC-H 9        | 0.113636538386  | 23   | 0.121970623732  | 25   | +7%        |                          0.0000 |
| TPC-H 10       | 0.216463521123  | 44   | 0.217899993062  | 44   | +1%        |                          0.0378 |
| TPC-H 11       | 2.28297924995   | 60   | 3.03103637695   | 60   | +33%       |     (run time too short) 0.0000 |
| TPC-H 12       | 0.699244916439  | 60   | 0.761795580387  | 60   | +9%        |                          0.0000 |
| TPC-H 13       | 0.157237693667  | 32   | 0.158344879746  | 32   | +1%        |                          0.3682 |
| TPC-H 14       | 1.7351616621    | 60   | 1.76626610756   | 60   | +2%        |     (run time too short) 0.0002 |
| TPC-H 15       | 1.3223618269    | 60   | 1.32369995117   | 60   | +0%        |     (run time too short) 0.6888 |
| TPC-H 16       | 0.699989378452  | 60   | 0.726635456085  | 60   | +4%        |                          0.0000 |
| TPC-H 17       | 0.071922019124  | 15   | 0.0755068510771 | 16   | +5%        |                          0.0000 |
| TPC-H 18       | 0.147115588188  | 30   | 0.171241939068  | 35   | +16%       |                          0.0000 |
| TPC-H 19       | 0.483642876148  | 60   | 0.48078635335   | 60   | -1%        |                          0.5958 |
| TPC-H 20       | 0.146263256669  | 30   | 0.147148549557  | 30   | +1%        |                          0.0003 |
| TPC-H 21       | 0.0334587022662 | 7    | 0.0333742797375 | 7    | -0%        |        (not enough runs) 0.5848 |
| TPC-H 22       | 0.900039494038  | 60   | 0.896938681602  | 60   | -0%        |                          0.2104 |
| geometric mean |                 |      |                 |      | +5%        |                                 |
+----------------+-----------------+------+-----------------+------+------------+---------------------------------+
```