@Bouncner 2019-04-28
Any ideas what’s the issue with Q14?