@mrzzzrm 2019-04-28
Yeah, https://github.com/hyrise/hyrise/pull/1647/commits/e9cfa7b722cf58a5a62c7ea7f8d97cf8985a2cdd fixed that. Re-running TPC-H right now.