@mrzzzrm 2019-04-28
Benchmark results in OP updated. Solid improvements for such a small deed.